requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==5.1.3
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.0
redis==5.2.0
sqlparse==0.5.2
typing_extensions==4.12.2
